Make rpc.js ESM compatible
